Update styles and move DOP page
