+++ components/common/top-nav/TopNav.module.css
@@ -265,3 +265,19 @@ a.dropDownItem:focus, a.dropDownItem:hover {
     background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16' fill='%2320b6b0'%3e%3cpath fill-rule='evenodd' d='M1.646 4.646a.5.5 0 0 1 .708 0L8 10.293l5.646-5.647a.5.5 0 0 1 .708.708l-6 6a.5.5 0 0 1-.708 0l-6-6a.5.5 0 0 1 0-.708z'/%3e%3c/svg%3e");
 }
 
+.arrow {
+    background-image: url("/images/chevron-right-teal.svg");
+    display: block;
+    transition: none !important;
+    background-repeat: no-repeat;
+    background-position: right 20px top 13px;
+}
+
+.arrow.active {
+    background-image: url("/images/chevron-right-grey.svg");
+    display: block;
+    transition: none !important;
+    background-repeat: no-repeat;
+    background-position: right 20px top 13px;
+}
+
