+++ components/data-oriented-programming/code/java-bbe/flexible-typing.md
