+++ components/data-oriented-programming/code/java-bbe/record-type.md
