+++ components/data-oriented-programming/code/java-bbe/stream-type.md
