+++ components/data-oriented-programming/code/java-bbe/type-constraints.md
