+++ components/data-oriented-programming/code/java-bbe/xml-support.md
