+++ components/data-oriented-programming/intro/Intro.js
@@ -20,7 +20,7 @@ import * as React from 'react';
 import { Row, Col, Container } from 'react-bootstrap';
 import Image from 'next-image-export-optimizer';
 
-import { prefix } from '../../../../utils/prefix';
+import { prefix } from '../../../utils/prefix';
 import styles from './Intro.module.css';
 
 export default function Intro() {
